pyproject.toml
[project]
name = "ratball"
version = "0.1.0"
description = "Motion tracking and audio/visual telemetry platform for the RATBALL omni-treadmill system"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "loguru==0.7.3",
    "numpy>=1.24.4",
    "pyyaml>=6.0.2",
    "ruff==0.11.6",
    "sparkfun-qwiic-i2c>=2.0.0",
    "sparkfun-qwiic-otos>=2.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.11.6",
]
